+++ lib/eve_online/esi/universe_category.rb
@@ -1,8 +1,12 @@
 # frozen_string_literal: true
 
+require 'forwardable'
+
 module EveOnline
   module ESI
     class UniverseCategory < Base
+      extend Forwardable
+
       API_ENDPOINT = 'https://esi.evetech.net/v1/universe/categories/%<category_id>s/?datasource=%<datasource>s'
 
       attr_reader :id
@@ -11,6 +15,13 @@ module EveOnline
         super
 
         @id = options.fetch(:id)
+      end
+
+      def_delegators :model, :as_json, :category_id, :name, :published,
+                     :group_ids
+
+      def model
+        @model ||= Models::Category.new(response)
       end
 
       def scope; end